+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Display chat in dashboard during a jitsi live
 - Use chat when a live is publicly available
 
+### Fixed
+
+- Prevent to display a deleted video
+
 ## [3.19.0] - 2021-05-10
 
 ### Added